split flask app
