+++ link_builder.py
@@ -32,6 +32,7 @@ class UrlBuilder(ABC):
     @abstractmethod
     def return_data(self):
         pass
+
 #Builder Concrete Classes
 class AdUrlBuilder(UrlBuilder):
     def __init__(self):
@@ -48,8 +49,8 @@ class AdUrlBuilder(UrlBuilder):
         now = datetime.now()
         self.hour = str(time(now.hour,now.minute,now.second))
     def generate_token(self):
-        AdUrlBuilder.get_date()
-        AdUrlBuilder.get_hour()
+        AdUrlBuilder.get_date(self)
+        AdUrlBuilder.get_hour(self)
         strings = string.ascii_lowercase + string.ascii_uppercase
         token = ""
         for x in range(7):
@@ -57,6 +58,7 @@ class AdUrlBuilder(UrlBuilder):
         self.link.token = token
     def return_data(self):
        return self.link
+
 class WithoutAdUrlBuilder(UrlBuilder):
     def __init__(self):
         self.link = Url()
