+++ run.py
@@ -1,38 +1,17 @@
-from flask import Flask, redirect, request, render_template, abort, jsonify
 import link_builder
-from flask_sqlalchemy import SQLAlchemy
-app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
-db = SQLAlchemy(app)
-#Director Class
-global domain_name
-#Change domain_name with your domain name
-domain_name = "127.0.0.1:5000"
-class UrlDatabaseModel(db.Model):
-    id = db.Column(db.Integer,primary_key=True)
-    real_link = db.Column(db.String(100),unique=False,nullable=False)
-    ip_addy = db.Column(db.String(20),unique=False,nullable=False)
-    ad_type = db.Column(db.String(5),unique=False,nullable=False)
-    token = db.Column(db.String(10),unique=True,nullable=False)
-    date = db.Column(db.String(20),unique=False,nullable=False)
-    hour = db.Column(db.String(20),unique=False,nullable=False)
-    def __repr__(self):
-        return f"Url('{self.id}','{self.real_link}','{self.ip_addy}','{self.ad_type}','{self.token}','{self.date}','{self.hour}')"
-class CreateUrl():
-    def __init__(self, ad_type):
-        self.ad_type = ad_type
-    def create(self):
-        if self.ad_type == "False":
-            builder = link_builder.WithoutAdUrlBuilder()
-            builder.get_ad_type()
-            return builder
-        elif self.ad_type == "True":
-            builder = link_builder.AdUrlBuilder()
-            builder.get_ad_type()
-            return builder
+from app import app, db
+from app import UrlDatabaseModel
+import link_builder
+from flask import redirect, request, render_template, abort, jsonify
+from app import CreateUrl
+
 @app.route('/')
 def index():
     return render_template("index.html")
+
+global domain_name 
+domain_name = 'http://localhost:5000'
+
 @app.route('/shortit',methods=["POST"])
 def short_it():
     #Building url
@@ -44,6 +23,7 @@ def short_it():
     db.session.add(created_url)
     db.session.commit()
     return redirect("/")
+
 @app.route('/api/shortit',methods=["POST"])
 def api_short_it():
     try:
@@ -63,10 +43,12 @@ def api_short_it():
             "Status": "Failure"
         }
     return jsonify(return_data)
+
 @app.route("/shortenedlinks")
 def shortenedlinks():
     URLs = UrlDatabaseModel.query.all()
     return render_template("shortenedlinks.html",URLs=URLs)
+
 @app.route('/<token>')
 def redirect_to(token):
     URLs = UrlDatabaseModel.query.all()
@@ -78,5 +60,7 @@ def redirect_to(token):
                 return redirect(("http://" + URL.real_link))
     print(token)
     return abort(404)
+
+
 if __name__=='__main__':
     app.run(debug=True)
